Add capability to restrict GBBEPx to certain dates
A new variable, gbbepx_days, is added that will set the dates allowed
for GBBEPx files. The variable uses delayed expansion, so you can
enclose with single quotes if you want to define in terms of another
variable (like any of the PDY variables). This functionality is useful
for development, when we want to restrict files to those we expect to be
available on-time for an operational run.

The default is to look at the current day and two previous days:
export gbbepx_days='$PDY $PDYm1 $PDYm2'