Fix problem with overriding gbbepx_pattern
Previous change to split GBBEPX_DATA_DIR into a TODAY and YESTERDAY
introduced a problem where gbbepx_pattern would have to be changed
depending on cycle and possibly not even work properly if you
wanted to override the default. Now, instead of changing the
gbbepx_pattern depending on the cycle, GBBEPX_DATA_DIR is set to
either the today or yesterday version and gbbepx_pattern can remain
unchanged.