Reenable reading headers & new format for WFABBA files
At some point, code that read the header of WFABBA files was
commented out, as well as the determination whether the data
included new fields or not. Presumably this was because the headers
were being stripped out in development.

However, when the WFABBA data started being populated by NCO on DCOM,
the data contained the headers as well as the new data fields. This
caused an error when attempting to read. Restoring the commented code
allowed the data to now be read from DCOM.